gcf_polar_to_gcs/requirements.txt
cloudevents==1.9.0
functions_framework==3.4.0
polar_accesslink==0.0.5
protobuf==4.24.1
xmltodict==0.12.0
google-cloud-storage==2.6.0
google-cloud-logging==3.6.0